Refactor a few casting checks. BCADREN: Also minor refactoring on a Kiku Passive.
Because MP is decremented before your_spells is called for regular
casting, the error messaging in the targeter was showing low MP messages
in wrong cases. This handles some casting-general checks differently
than other uselessness checks so that they can be consistently performed
in cast_a_spell, and therefore ignored within your_spells. It also
removes one existing messy unwind case that worked around this. The
situation is still a bit of a mess though.