Add build test for headers with fixed includes
This helps verify that headers are including everything they use (IWYU),
since if they are not, compilation will fail. We don't actually do
anything with the compiled objects.

This complements implicit checks for cc files, which must include their
dependencies because they could not successfully compile otherwise.

This commit only includes a subset of all header files. Header files can
be appended to this list as they are fixed.