New monster: bunyip
Bunyips are an Australian cryptid of unclear description. This
implementation of them treats them as a sort of lesser juggernaut,
hitting for 40 three times. (Their defenses are substantially
weaker than the real juggernaut.) They also have howler monkeys'
warning cry, which they can waste some time with to give players
more of a chance to kill them or flee.