Generate temple before dungeon
This moves Temple before D in the pregen order, and pregens it during
setup for normal games. The level is fully decided at this point, and it
shouldn't interact with the rng for any other level, so order
effectively doesn't matter anyways. This is a long-requested feature to
avoid the popup where entering temple forced pregeneration of the rest
of the dungeon.