Improve Text wrap recalculation checks
This function is *extremely* slow for large text boxes on local tiles,
and this check eliminates some level of re-calling when doing initial
widget sizing. I think there's still something wrong with this call
somewhere deeper in that I haven't figured out, but this helps somewhat
at a global level.