Fix some veto-triggering issues
This is mostly a matter of adding `transparent` to more vaults, but in
some cases I found issues with the vaults themselves that led to
disconnected areas needing a tweak to some cells or some explicit opaque
masking. Hopefully the additional `transparent` tags don't lead to
further issues, but as long as the vault genuinely lacks disconnected
regions, as far as I can tell it is always better to have it be tagged
transparent, even for vaults with a small number of exits.