Improve layout_gridlike's handling of water/lava (Skrybe)
This fixes a couple of bugs at once:
 * layout_gridlike was forcing vetos if it generated shallow water
   closed rooms, because they registered as disconnected areas.
 * closed rooms acted as tele closets for species that could inhabit the
   terrain.

It might be better to do this on a bit more of a case-by-case basis to
avoid placing no_tele_into in some corner cases (outer material plants,
inner material shallow water). However, I think that overkill is better
than underkill, and a comment by @Skrybe in the issue thread provided a
ready-made solution.

Resolves #1234. Also mantis 7563.