Despoiler /poly (gammafunk)
Pre-choose a set of 3 monsters that any given monster can turn into
(filtering by holiness & habitat and weighting by HD and demon tier),
and display them on the /polymorph targeter. This should make it much
more possible to make tactical choices about the use of /polymorph.

This does not affect other sources of polymorphing, such as Xom or Nem.