Reorder, fix checks in _valid_type_morph
Fix some old issues exposed by recent poly changes. It's possible that
one of the many things in this disjunction would have checked for
MONS_NO_MONSTER, but there was no direct check for it. In addition,
checking holiness first is generally a recipe for crashes because it is
a complicated check that presupposes valid monster data. In this case,
this call guaranteed to crash on MONS_NO_MONSTER, so firing newpoly on
such a monster would crash instead of show the "momentary change"
message.