Don't set greedy_place for items in hostile areas
As things stood, if autopickup was off and there was an unexplored space
behind a runed door, _explore_find_target_square() called pathfind()
once to no avail, called it a second time with try_fallback set, and
made that the target.

If autopickup was on, it called pathfind() once, cleared it (because it
had a negative travel_point_distance[] value), and printed a "Partly
explored" message.

This patch prevents greedy_place from being set if
travel_point_distance[x][y] is negative, so the square containing the
item can become the target.