Tweaks to quickstart.md
This is shown in game, so needs to look reasonable as text. Adjust the 
headers and a few other details.  It still uses a bit too much
formatting for this, so it would be ideal if we had a way of stripping
out the md and generated a .txt.

Also, remove food, and a few other minor wording tweaks.