Remove skill training restrictions (hellmonk)
After the old XP pool system was removed and the current skill training
system was added, one odd edge case was that you had to have an item, spell,
or god ability somehow relevant to a skill to be able to train it. This was
not particularly consistent. For example, training got special-cased to be
always allowed to avoid annoyance when you'd thrown everything you had, and
pain brand, though depending on necromancy, did not allow you to train it.

In general, it's better to have less complexity and fewer special cases, so
by default it seems good to remove these rather complex restrictions. The
argument against has been that training these skills is rarely useful (true,
but so are many things players do!), and that a massive skill screen would
intimidate new players. Thankfully, to avoid the latter case, we can retain
much of the old code to hide many skills by default, while still allowing
experienced players to use them, in the same way that they can use manual
training instead of automatic. In fact, since we no longer need to special
case Throwing, new players will often see a smaller skill screen than before!