First-pass implementation for quivering abilities
There's a lot of idiosyncracy in the ability list, but this commit gets
the main points and works for everything I've tested directly. Probably
further tweaking is required for some special cases, and this may
require some iteration with actual player testing to really shake out
the use cases. (And the ability code is a mess, a lot more stuff could
be dataified...)

Main TODO: non-targeted abilities use the "untargeted cast" interface
when [f]iring from the quiver, but several of these would benefit from a
static targeter. (For the ones that affect the player, which is most of
them, there's not much difference, but these could still use the
self-targeted static targeter when firing.)