Change the behavior of the CLUA_BINDINGS #define
This commit changes things so that what gets compiled out by removing
this define is only the ability to run clua from the rc file and macros.
Default clua bindings in dat/clua are called, as are all hooks etc.
(Most will be unable to do anything, but this would allow someone to set
a default value for any hooks in dat/clua code.)

*Motivation.* The main use case for this that I can think of is
security: a security-conscious server admin may prefer their users to
not be able to run clua in their rc file. However, this use case has
been severely broken for some time, because the following were also
disabled by undefining this variable: autofight, stash tracker
annotation, delay interruption, kill annotation, and a few other minor
things.  Realistically no one is going to want to play modern crawl
without at least the first two of these.  Because of the (potential)
security application, I don't want to remove it entirely, but if it is
going to be usable for that, it needed some work.

In addition, the existence of this #define required anyone implementing
a clua binding to get the preprocessor setup right, and if they really
wanted to do that, to test on two builds (one of which required editing
the Makefile). Nearly every dev comment I can find about CLUA_BINDINGS
going back to 2011 is basically asking why it exists. So this commit
is an attempt to provide some value.

One awkwardness: there is a bunch of lua code in
dat/defaults/autopickup_exceptions.txt. Because this is treated as an rc
file, it will error (and not work) on a non-user-clua build. This is all
a bit hypothetical now but possibly it would be better for this code to
be implemented in a dat/clua file rather than a default rc include.