Refactor and consolidate melee attacks
This commit refactors melee attacks, including CMD_ATTACK_... and
reaching attacks, into a single action type. This action supports
targeting not just for reaching attacks but for regular melee attacks,
leading to a more uniform behavior (as well as allowing you to examine
enemies directly from an attack interface for all attack types). This
class supports all the bells and whistles of other action commands, e.g.
it will do noninteractive attacking if a target is pre-chosen. It's
currently not used in a quiver at all, but created on the fly as part of
an attack.

For backwards compatibility I have left the behavior of
CMD_EVOKE_WIELDED intact, but `v` is now bound to a new
CMD_PRIMARY_ATTACK that activates targeting for any wielded weapon, be
it regular, reaching, or ranged. Historically, there were various types
of things triggered with this key while wielded, but more recently `v`
is used almost entirely for reaching except for some special cases: two
artefacts (staff of olgreb and staff of dispater) as well as unusual
techniques like wielding a wand. Since all of these special cases now
can be accomplished via the quiver, it doesn't seem so necessary to keep
to the old scheme as a default. There's still unfortunately a bit of
scaffolding in the evoke code because of this compatibility issue, but
it's at least less intrusive.

Additional notes:
* targeter_cleave is upgraded to support changing aim, and used for
  `v` on axes.
* Rift targeting is simplified and now correctly shows range.
* Bump attacking is unchanged, i.e. doesn't use this code path.
* If someone's really bothered by having to quiver staff of olgreb's
  OTR ability to trigger it, they can rebind `v` back to
  CMD_EVOKE_WIELDED. `V` will also continue to work.
* It is still possible to use `V` to evoke a reaching weapon.