Fix an autofight / throw interaction (vt)
This simplifies this code path to only use the targeting in throw_it,
rather than adding an extra layer of targeting around
CMD_THROW_ITEM_NO_QUIVER. This layer was here for historical reasons and
everything seems to work fine without it. The specific bug this is
targeting is that because the action treated this as a case of
non-interactive targeting it would apply the autofight threshold.

This commit also fixes a bug where you could use `f` to fire ammo into
walls with portal projectile, probably causing a crash.