Show inert evokers in `V` menu (Lici)
This restores previous behavior, and marks inert evokers as dark grey in
inventory menus in general (basically so that they'll show up that way
in `V`).  Not sure if there are other things that could be marked dark
grey following this, or if there's a convenient way to do that. But this
makes the appearance of `V` match the quiver menu.