Don't show static targeters for `Z`
People seemed confused by this, so I'm removing them for now. I can't
tell whether this confusion is because of change or intrinsic confusion.
It's true that many static targeters are extremely boring (e.g.
self-targeted ones and summon ones that just show range 2). I'm leaving
them for the fire interface, because even though I've also seen this
described as "weird" it's pretty important for consistency, and if you
want to skip the targeter you can just use `p` or shift-tab in this
mode. (Also, I do really think these targeters are extremely useful for
new players.)

What might be best is if z/Z both had a toggle that let you choose to
see the static targeter if you want to check range.