Set you.on_current_level to false during load
This has the effect of suppressing certain UI things, and keeping
quivers from crashing in case UI updates during load force webtiles to
call send_player early. This was in particular happening for skill menu
on load cases.

I've also adjusted where the skill menu is called from in these early
game cases; this didn't prevent the crash or anything but it seemed
wildly odd to open the skill menu when the game is in the middle of
restoring.