Better handle invalid lockfiles
After CBR2 lost power the other day, it left a bunch of empty lockfiles,
which aren't being handled very well. This will simply clear a lockfile
if it is empty or otherwise contains something not convertable to an
int. (The case where it contains a stale process number is already
handled.)