Do some caching for spell quiver enabled checks
For beams (and especially explosions) this is a pretty heavy call,
because it checks if there are enemies in range, and it isn't ideal that
it gets called every time the quiver gets redrawn; I've already fixed
some cases where this call was leading to noticeable slowdowns.
Moreover, this call has been revealing a lot of cases where a redraw (or
webtiles _send_player) happens as a side effect at a time where there is
some invalid state for calculating nearby enemies, causing crashes. I've
been able to fix up some of these on a case by case basis, but others I
haven't even been able to replicate (e.g.
https://webzook.net/soup/morgue/trunk/eld/crash-eld-20201128-170107.txt,
which seems to involve a more() triggered in the middle of monster_die
on vehumet MP gain, where there is a partially dead monster at that
point in the call, triggered by some kind of complex fireball+inner
flame scenario).

So, for both of these reasons, this commit changes things so that the
spell quivered enabled status is only updated once per input cycle. This
could be extended to other quiver actions needed but right now as far as
I know (beam) spells are the ones that are causing issues.