Re-simplify quivering effect of wielding a launcher
b220303a8 tried to improve autoquivering on wield but instead (for most
people) just made it more confusing.  After this commit, by default
wielding a launcher will always change the main quiver as well,
regardless of what is quivered. This works reasonably well with the
other change in b220303a8 that switches to the last used quiver on
quiver invalidation, so switching away from the launcher (except for
slings+stones) usually does something intuitive. This behavior (in
contrast to the behavior before b220303a8) can now be disabled entirely;
if you set `launcher_autoquiver = false` wielding a launcher will never
set the quiver. This will be useful at least for characters that are
using both launchers and spells.