Fix autofight_fire_stop
Broken because the lua code was using the wrong variable name. This
commit restores the previous default behavior of shift-tab/p acting like
`ff` instead of `f.`, lets this be configured again in the rc via
`autofight_fire_stop = true`, and also has all quiver actions
consistently respect this option.  I've gotten some feedback from
players suggesting that this might actually be better with a default of
false (from players interpreting the bug as a new default), but for now
I'll leave it as it was before the quiver rewrite.