Various tweaks to monster and spell descriptions
This fixes a few apparent mistakes, and more cleanly aligns fields in
both of these description popups. (The unifying factor is the internal
use of `_print_bar`.)

This commit involves a rather alarming nbsp replace on the javascript
side to get these tweaks to render in webtiles, but on the other hand it
does seem very plausible to me that any leading spaces in one of these
popups will be intentional.