Provide a way to use pkg-config on mac
This commit adds a build setting `FORCE_PKGCONFIG` that if defined,
overrides the default `NO_PKGCONFIG` setting for mac builds.
Previously, doing something like this required editing the Makefile.
This setting should be useful for things like homebrew that repackage
crawl and patch the Makefile (see #1367). It may also be helpful for
people trying to build on new apple ARM devices where our contrib
version can't be updated far enough to provide direct support (though
this is a bit speculative).

A commit in #1367 suggests just always using pkg-config if available,
but I don't think this is yet a good idea on mac. Because Apple doesn't
itself provide it, there are at least 4 distinct ways that I know of to
(often accidentally) get pkg-config installed, none of them are very
standard or compatible, it is possible to get more than one version
which interact badly (e.g. if you have both macports and anaconda, or
homebrew and anaconda), and in my past testing they are just quite
unreliable overall, probably exactly because they aren't widely used. In
testing this current commit, I do seem to have succeeded in getting a
succesful pkg-config build, but this is the first time.  The contribs
for mac builds, on the other hand, are extremely well-tested and
reliable, and not at all sensitive to the various package managers that
a mac user could have installed. So I think until this situation becomes
a bit more stable (if ever), it is better to default to NO_PKGCONFIG and
require the user to be aware that they are taking action to override
using contribs.