Fix webtiles tooltips for statuses with counts
Statuses like "Flay (-7)" and "Zot (999)" had their tooltips broken,
because the lookup was using this full string. Instead, split at the
space. This commit also adds some text when there is no description so
that it is more apparent that there should be a tooltip (before, a tiny
empty box was displayed).