Refactor your_talents somewhat
Previously, to check if the player had an ability, you needed to
construct the entire talents list and see if the ability is there. This
separates out code that checks whether a player has an ability from
your_talents so that it can be directly called in a less overkill
fashion. There's further refactoring along these lines that would be
possible, for god abilities in particular.

This commit also tweaks some flight logic in a way that hopefully
doesn't break it, but it seemed to be subtly wrong before.