Separate launcher from throwing actions
The main reason for this is so to differentiate throwing stones from
launching stones, leading to better behavior on switching between a
sling and melee weapon. It has the slightly odd result that you can
quiver stones for throwing while wielding a launcher, which used to be
impsosible, but this seems mostly harmless?

This commit also removes the ranged side of auto_switch. As far as I can
tell, this only ever worked when the quiver was empty, making it almost
entirely useless even before the quiver improvements to launching. Now,
there's always a way to fire a wielded launcher regardless of quiver
state. If someone can tell me there's a use case I am missing I can try
to restore it. This option still works as before for melee weapons.