Do basic deduplication when rendering tile overlays
This is very minimal: it just prevents stacking completely adjacent
identical overlays. This is intended to handle cases with large numbers
of bennus (e.g. in a zig) where many, many explosions can get stacked on
top of each other; in these cases, the number of overlays was easily
reaching the cap (20) which causes a crash. Since layering these
overlays does nothing, simply prevent it. If this system is ever used to
layer overlays with alpha values, this may have unexpected effects, but
to my knowledge it isn't.