Fix a bunch of local tiles clicking stuff
We got a report that e42c0a684a0 broke click-to-throw. I'm still not
sure what this report was about, but on looking into this, I found any
number of things broken about clicking-related stuff. Some of this is
related to quivers, some is older. This commit fixes at least the
following bugs:

* The tooltip info was a mess and often wrong
* Some of the conditions for whether to report "Travel" tooltips were
  out of sync with the travel code
* open door tooltips were being calculated wrong
* quivers were being triggered, but by passing key commands to the fire
  interface, rather tha by setting a target directly
* On shift-click to cast with too low mana, a bunch of key errors were
  printed
* primary attacks with launchers weren't handled correctly, and only
  triggered movement